backend/requirements.txt
requests
pandas
python-dotenv
openpyxl
